Add progress bar to the product list screen
Added a CircularProgressIndicator to the product_list_fragment.xml file.
Progress bar is shown while product list is being fetched. It's hidden if
the network connectivity is lost before the product list is retrieved.
In addition, network connectivity warning is now only shown if the connectivity
is lost before the list is retrieved. If the user taps on an item while the
there is no connection, an alert dialog will pop up in the future.