Implement product list sorting
+Added ProductSortBottomSheetFragment.kt and product_sort_bottom_sheet.xml
The bottom sheet is accessed by pressing the new SortProductButton in the
ProductListFragment. Using a button inside the bottom sheet calls the
onBottomSheetButtonClick method of ProductListFragment with the new
ProductType enum as a parameter.
+GetProductsUseCase.kt is changed to take the ProductSortType as a parameter,
which is in turn given to the ProductRepository.kt as a parameter.
+Added new methods to the YamlApi.kt for making API calls that return sorted
product list. ProductRepositoryImpl.kt's getProductList was changed to call
a different YamlApi.kt method depending on the sort type.
+Added freight_sans_bold.otf font, currently used by the sort product button
+Added bottom sheet related styles