Add ProductDetail related xml files and recyclerview adapters
+ Implemented view components in product_detail_fragment.xml
+ Added button list items for size and color buttons
+ Added adapter object for recyclerviews that contain color and size buttons
These adapters will be assigned to their recycler views from the fragment.
The adapters change the button stroke colors when they are clicked to indicate that
they are selected. Adapters also require a callback listener object as a parameter
in order for fragment to apply changes to the UI when a button is clicked.